Refresh default services using viper.UnmarshalKey
For some reason Viper is behaving differently when unmarshaling the
entire configuration file in comparison with just one key.
In the first case it returns an error, and in the second one
(unmarshaling one key), it does not fail and returns the proper configuration.

In this case, the elasticsearch environment variables contains dots,
that's why they are not properly read.