Merge a user-defined configuration with default values
The user-defined configuration could be located at the path of execution
of the CLI, so that it will merge this one with the default.

The rationale under this aaproach is to allow users to configure the tool
with their own configurations and values, being possible to share
with other team members any valuable configuration.

If seen valuable for the entire team, we could merge it to the defaults.