(#42) Use gobuffalo's packr2 to package static resources into the binary
It will allow docker compose to inspect the files and be able to execute
a docker-compose file from it.

We are going to write the bytes of the in-memory compose file to a
temporary file (at os.Tmpdir), to be able to read it. We will remove it
in a deferred way