feat: support persisting the state of each run in a file
A run is an execution of the tool, be it a stack or a service. With the
tool we can add or remove services to a run

The .run file will contain the stack (optional) and the services in the
run, including the environment used for the ran, which will allow to
restore that state for future updates in the run.

We also support recovering and desttroying the state for a run