feat: support adding the environment variables coming from the supported-versions.yml file
This is important for variants, as they come with custom environment variables
in the YML file. For that reason we are generating a struct on-the-fly
to parse the incoming YML file, and then we loop through the keys to
autodiscover the custom environment variables.
Finally, we merge the custom environment variables with service's defaults