feat: start translation of check_parity for beats
We are checking the number of files monitoried by filebeat, obtained
from the different document types present in the _source document.

We use Jeffail/gabs to get a Container struct on-the-fly when reading a
JSON file. We will be able to search through this container using paths
with the '.' notation (i.e. 'hits.hits._source.type')