chore: use Observability's Kibana image until a bug is fixed
It seems that kibana is no handling properly the version of the latest
agent when this is enrolled. This 3-day old kibana version supports it.

A nice improvement would come if we have the different kibana versions,
so we are able to configure the environment to run it with that version
of the stack, instead of using 8.0.0-SNAPSHOT.

Using moving targets it's always a problem