feat: send errors for merges to Slack channels, depending on the suite
We will send the errors to the observability-robots channel by default,
being a responsibility of the scheduled pipelines to configure the proper
Slack channel.

We will send Helm and Integrations suite errors to the #integrations
channel, and the Ingest-Manager errors to the #ingest-management one.

This behavior will affect only to merges to master, not to PRs