Kubernetes autodiscover suite (#1064)
Adds a new suite to test Kubernetes autodiscover features.
Reproduces some issues with short-living containers affecting Beats before 7.13
and for what we didn't have tests coverage.

It allows to define the scenarios using the gherkin language for feature files.
Scenarios use templates that are mostly kubernetes manifests. A simple system of
options is included to allow to select configuration blocks in these templates.

It uses kubectl to interact with a kubernetes cluster. If no cluster is available, it
starts one with kind. Each scenario is executed in a different namespace, namespace
is destroyed after the scenario is run to clean all created resources.

It currently works only with loads in a single namespace, in a single node, this is
enough to test many autodiscover cases, but it could be extended in the future to cover
multi-node or multi-namespace scenarios.

Co-authored-by: Manuel de la Peña <mdelapenya@gmail.com>
Co-authored-by: Adam Stokes <51892+adam-stokes@users.noreply.github.com>
(cherry picked from commit c52395dc53adf8ed95ff0187de82c86498d0abdc)