feat: add an elastic-package provider
First, it will only support the Fleet test suite, exiting with an error
if it's another test suite.

Then, it will use elastic-package to provision the runtime dependencies
of the Fleet test suite, running the stack withouth the elastic-agent
service. This service will be provisioned using testcontainers directly,
using the existing base images: centos/debian with systemd and the oficial
Docker image for the elastic-agent