Move OS specific commands into actions
This adds support for what I am calling ServiceOperatorAction. We are moving
specific executions into this namespace so that depending on the action we want
to run it will do the right thing on the underlying OS.

This is part 1 of the work to pave the way for supporting Windows with our first
action on checking process state.

Signed-off-by: Adam Stokes <51892+adam-stokes@users.noreply.github.com>