Move OS specific commands into actions (#1519)
* Move OS specific commands into actions

This adds support for what I am calling ServiceOperatorAction. We are moving
specific executions into this namespace so that depending on the action we want
to run it will do the right thing on the underlying OS.

This is part 1 of the work to pave the way for supporting Windows with our first
action on checking process state.

Signed-off-by: Adam Stokes <51892+adam-stokes@users.noreply.github.com>

* use type assertion for action opts

This lets us generalize the options passed to an action since each action will
already have a known type to handle any additional options needed for the action
to perform.

Signed-off-by: Adam Stokes <51892+adam-stokes@users.noreply.github.com>

* Minor cleanups

Signed-off-by: Adam Stokes <51892+adam-stokes@users.noreply.github.com>
(cherry picked from commit 4fce879df6d776caf98cd5452953c8c25f564d3e)