feat: support bootstrapping the stack with a new kibana config
The bootstrap method is called before the test suite (only once) AND before
each test scenario, as it has been defined as a Background clause. If an
scenario wants to configure a new config file for Kibana, a new profile file
with the new kibana config must be created at the compose level, using
the ProfileName as directory name.