chore: rename parameter to avoid collitions
The DEVELOPER_MODE variable could affect the variable used by the project
to avoid uninstalling/unenrolling the elastic-agent after a test scenario
finishes.

We have identified that the Jenkins params are automatically converted
into environment variables, and that could affect the aforementioned
variable.

We have created two PRs with same codebase, but defining the DEVELOPER_MODE
Jenkins param with true and false values:
- true: the instances are kept and the build consistently finishes without SSH disconnections
- false: the instances are destroyed and the build consistently fails with SSH disconnections