pkg/downloads: make sure nil resolvers are interpretted correctly
Previously, the returned values from NewArtifactURLResolver could be a
nil *ArtifactURLResolver. This was incorrectly being compared to nil in
the context of a DownloadURLResolver in getDownloadURLFromResolvers and
so being ignored, finally resulting in a nil pointer dereference when
the Resolve method was called. The change here makes the nil returned by
NewArtifactURLResolver be a nil DownloadURLResolver which will be
recognised in getDownloadURLFromResolvers and so not used.